Learn records, streams

+++ src/homeworks/dmytro_k/hw_2023/hw_01_10_23/task1/StationaryExtractor.java
@@ -10,7 +10,7 @@ public class StationaryExtractor extends Extractor {
 
     private static final ExtractorType EXTRACTOR_TYPE = ExtractorType.STATIONARY;
 
-    public StationaryExtractor(int expectedPerformance, int quantity) {
+    public StationaryExtractor(double expectedPerformance, double quantity) {
         super(EXTRACTOR_TYPE, expectedPerformance, quantity);
     }
 }
